requirements.txt
# Core
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Scientific
scipy>=1.10.0
scikit-learn>=1.3.0

# Database
sqlite3 (built-in)

# Scheduling
apscheduler>=3.10.0

# Configuration
pyyaml>=6.0

# Logging
python-dotenv>=1.0.0

# Utils
tqdm>=4.65.0
pandas>=2.0.0
